requirements.txt
pandas
scikit-learn
mlflow
numpy
flask
gunicorn
requests
flake8
pytest          
python-dotenv    